Mauricio dev

+++ easyturn/src/main/java/com/backend/easyturn/services/AdministratorService.java
@@ -1,12 +1,15 @@
 package com.backend.easyturn.services;
 
 import com.backend.easyturn.entities.Administrator;
-import com.backend.easyturn.entities.Appointment;
+import com.backend.easyturn.entities.DTOs.AdministratorDTO;
+import com.backend.easyturn.exceptions.AppException;
 import com.backend.easyturn.repositories.AdministratorRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.List;
 
 @Service
@@ -15,21 +18,75 @@ public class AdministratorService {
     @Autowired
     private AdministratorRepository administratorRepository;
 
-    public int createAdministrator(Administrator administrator){
-        Administrator adminCreated = administratorRepository.save(administrator);
-        return adminCreated.getIdAdministrator();
-    }
-    public void deleteAdministrator(int idAdministrator) {
-        this.administratorRepository.deleteById(idAdministrator);
+    public AdministratorDTO createAdministrator(Administrator administrator){
+        try{
+            Administrator adminFound = this.administratorRepository.findByMail(administrator.getMail());
+            if(adminFound != null){
+                throw new AppException("Este mail ya existe!", HttpStatus.CONFLICT);
+            }
+            Administrator adminCreated = administratorRepository.save(administrator);
+            AdministratorDTO adminDTO = new AdministratorDTO(adminCreated.getId(), adminCreated.getMail(), adminCreated.getName(), adminCreated.getLastName());
+            return adminDTO;
+        } catch (Exception e) {
+            throw new AppException(e.getMessage(), HttpStatus.BAD_REQUEST);
+        }
     }
 
-    public Administrator getAdministrator(int idAdministrator) {
-        return this.administratorRepository.findById(idAdministrator).get();
+
+    public void deleteAdministrator(int idAdministrator) {
+        try {
+            if(!this.administratorRepository.existsById(idAdministrator)){
+                throw new AppException("El administrador no existe!", HttpStatus.NOT_FOUND);
+            }
+            this.administratorRepository.deleteById(idAdministrator);
+        } catch (Exception e) {
+            throw new AppException(e.getMessage(), HttpStatus.CONFLICT);
+        }
     }
-    public List<Administrator> getAllAdministrator() {
-        return this.administratorRepository.findAll();
+
+    public AdministratorDTO getAdministrator(int idAdministrator) {
+        try{
+            Administrator administrator = this.administratorRepository.findById(idAdministrator)
+                    .orElseThrow(() -> new AppException("El administrador no existe!", HttpStatus.NOT_FOUND));
+            AdministratorDTO administratorDTO = new AdministratorDTO(administrator.getId(), administrator.getMail(), administrator.getName(), administrator.getLastName());
+            return administratorDTO;
+        } catch (Exception e) {
+            throw new AppException(e.getMessage(), HttpStatus.CONFLICT);
+        }
     }
-    public Administrator updateAdministrator (Administrator administrator) {
-        return this.administratorRepository.save(administrator);
-    } //creo que reemplaza el existente por id --> ver porque no estoy seguro (otra opcion es traerlo por id, setear los datos nuevos y volverlo a guardar)
+
+    public List<AdministratorDTO> getAllAdministrator() {
+        try{
+            List<AdministratorDTO> administratorDTOS = new ArrayList<>();
+            List<Administrator> administrators = this.administratorRepository.findAll();
+            if(administrators.isEmpty()){
+                throw new AppException("No hay administradores cargados!", HttpStatus.NOT_FOUND);
+            }
+            for(Administrator admin:administrators){
+                AdministratorDTO adminDTO = new AdministratorDTO(admin.getId(), admin.getMail(), admin.getName(), admin.getLastName());
+                administratorDTOS.add(adminDTO);
+            }
+            return administratorDTOS;
+        } catch (Exception e) {
+            throw new AppException(e.getMessage(), HttpStatus.CONFLICT);
+        }
+    }
+  
+    public AdministratorDTO updateAdministrator (Administrator administrator) {
+        try{
+            Administrator adminFound = this.administratorRepository.findByMail(administrator.getMail());
+            if(adminFound != null){
+                throw new AppException("Este mail ya existe!", HttpStatus.CONFLICT);
+            }
+            Administrator admin = this.administratorRepository.findById(administrator.getId())
+                    .orElseThrow(() -> new AppException("El administrador no existe!", HttpStatus.NOT_FOUND));
+            admin.setMail(administrator.getMail());
+            admin.setPassword(administrator.getPassword());
+            Administrator administratorModified = this.administratorRepository.save(admin);
+            AdministratorDTO administratorDTO = new AdministratorDTO(administratorModified.getId(), administratorModified.getMail(), administratorModified.getName(), administratorModified.getLastName());
+            return administratorDTO;
+        } catch (Exception e) {
+            throw new AppException(e.getMessage(), HttpStatus.CONFLICT);
+        }
+    }
 }
